+++ src/test/java/kitEditor/KitArchiveTest.java
@@ -14,7 +14,12 @@ class KitArchiveTest {
         tmpFile.deleteOnExit();
         Sample[] originalSamples = new Sample[3];
         originalSamples[0] = new Sample(new short[10], "FOO");
-        originalSamples[2] = Sample.createFromOriginalSamples(new short[10], "BAR", tmpFile, 1, false);
+        originalSamples[2] = Sample.createFromOriginalSamples(new short[10],
+                "BAR",
+                tmpFile,
+                1,
+                false,
+                true);
         KitArchive.save(originalSamples, tmpFile);
 
         Sample[] s = new Sample[3];
@@ -26,6 +31,7 @@ class KitArchiveTest {
         Assertions.assertEquals("BAR", s[2].getName());
         Assertions.assertEquals(10, s[2].lengthInSamples());
         Assertions.assertEquals(1, s[2].volumeDb());
+        Assertions.assertTrue(s[2].halfSpeed());
         Assertions.assertEquals(tmpFile.getAbsolutePath(), s[2].localPath());
     }
 }