+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Kit Editor: improved sound playback quality.
  - Kit Editor: click sample view to play.
  - Kit Editor: MPC-like UI with pads.
+ - Kit Editor: Half-speed setting now also affects "Add sample".
  - Palette Editor: improved mid-tone generation.
  - Various file dialog improvements.
  - Improved command line feedback.